Progress on the creator design
