+++ .gitignore
@@ -368,3 +368,6 @@ FodyWeavers.xsd
 
 # VS Code
 .vscode/
+
+.git.backup
+
