+++ Syndiesis/Controls/Editor/QuickInfo/ISymbolInlinesRootCreatorContainer.cs
@@ -1,0 +1,4 @@
+﻿namespace Syndiesis.Controls.Editor.QuickInfo;
+
+// Marker interface to avoid the generic explosion from the real base type
+public interface ISymbolInlinesRootCreatorContainer;
