+++ Syndiesis/Controls/Editor/QuickInfo/ISymbolItemInlinesCreator.cs
@@ -5,6 +5,17 @@ namespace Syndiesis.Controls.Editor.QuickInfo;
 
 public interface ISymbolItemInlinesCreator
 {
-    public abstract GroupedRunInlineCollection Create(ISymbol symbol);
+    public sealed GroupedRunInlineCollection Create(ISymbol symbol)
+    {
+        var inlines = new GroupedRunInlineCollection();
+        Create(symbol, inlines);
+        return inlines;
+    }
+
     public abstract void Create(ISymbol symbol, GroupedRunInlineCollection inlines);
+}
+
+public interface ISymbolItemInfoLanguageContainer
+{
+    
 }
